refactor: rename BackupDBClient to DocumentClient and update usages
